Make a Next.js test app login with a Central Dogma server (#759)
Motivation:

A dev server running on `3000` port redirects to the `http://DOGMA_HOST/link/auth/login` to get configured login page URL. The client expects the server to redirect to the dev server's login page if Shiro auth provider is configured. However, a Central Dogma server assumes the web client is hosted in the same origin and returns a relative path. As a result, the dev client loses a chance to render the login page and a Central Dogma hosted login page is rendered instead.
```
http://127.0.0.1:3000/ -> http://127.0.0.1:36462/link/auth/login ->
http://127.0.0.1:36462/web/auth/login
```
To request a login request from the dev server, the Central Dogma server should redirect to `http://127.0.0.1:3000/web/auth/login`.

Modifications:

- Prefix a host to the login path if `return_to` exists and redirect to the absolute URL.

Result:

A Next.js dev server can login with a Central Dogma server.